pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deploy-cli"
version = "0.1.0"
description = "CLI for triggering AWS CodePipeline executions by friendly alias"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Ranjeet Kumar" }]
dependencies = [
    "click>=8.1.7,<9",
    "rich>=13.7.0,<14",
    "boto3>=1.34.0,<2",
    "pydantic>=2.6.0,<3",
    "pyyaml>=6.0.1,<7",
    "questionary>=2.0.1,<3",
    "pyfiglet>=1.0.2,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-mock>=3.12,<4",
    "moto[codepipeline,sts]>=5.0,<6",
    "ruff>=0.4,<1",
]

[project.scripts]
aws-deploy = "deploy_cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["deploy_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
